switch to using u32s

+++ src/graphics/draw/line.rs
@@ -1,13 +1,13 @@
 use crate::{Color, Context, Point};
 
 pub fn draw(ctx: &mut Context, p0: Point, p1: Point, color: Color) {
-    let mut x0 = p0.x;
-    let mut x1 = p1.x;
-    let mut y0 = p0.y;
-    let mut y1 = p1.y;
+    let mut x0 = p0.x as f32;
+    let mut x1 = p1.x as f32;
+    let mut y0 = p0.y as f32;
+    let mut y1 = p1.y as f32;
 
-    let dx = (x0 - x1) as f32;
-    let dy = (y0 - y1) as f32;
+    let dx = x0 - x1;
+    let dy = y0 - y1;
 
     if dx.abs() > dy.abs() {
         if x0 > x1 {
@@ -16,10 +16,10 @@ pub fn draw(ctx: &mut Context, p0: Point, p1: Point, color: Color) {
         }
 
         let m = dy / dx;
-        let mut y = y0 as f32;
+        let mut y = y0;
 
-        for x in x0..=x1 {
-            ctx.texture.set_pixel(x as usize, y as usize, color);
+        for x in (x0 as usize)..=(x1 as usize) {
+            ctx.texture.set_pixel(x, y as usize, color);
             y += m;
         }
     } else {
@@ -29,10 +29,10 @@ pub fn draw(ctx: &mut Context, p0: Point, p1: Point, color: Color) {
         }
 
         let m = dx / dy;
-        let mut x = x0 as f32;
+        let mut x = x0;
 
-        for y in y0..=y1 {
-            ctx.texture.set_pixel(x as usize, y as usize, color);
+        for y in (y0 as usize)..=(y1 as usize) {
+            ctx.texture.set_pixel(x as usize, y, color);
             x += m;
         }
     }
